requirements.txt
discord.py
sentence_transformers
qdrant_client
python-dotenv
arq
watchfiles
openai
